+++ src/library/DIPS.Mobile.UI/Components/Searching/SearchBar.cs
@@ -7,8 +7,7 @@ namespace DIPS.Mobile.UI.Components.Searching
         public SearchBar()
         {
             this.SetAppThemeColor(IconsColorProperty, ColorName.color_icon_neutral);
-            // TODO: Use placeholder semantic color
-            this.SetAppThemeColor(TextColorProperty, ColorName.color_neutral_60);
+            this.SetAppThemeColor(TextColorProperty, ColorName.color_text_placeholder);
             this.SetAppThemeColor(iOSSearchFieldBackgroundColorProperty, ColorName.color_neutral_05);
 
             Unloaded += OnUnloaded;