+++ src/library/DIPS.Mobile.UI/Components/Selection/ISelectable.cs
@@ -5,7 +5,6 @@ namespace DIPS.Mobile.UI.Components.Selection;
 
 internal interface  ISelectable
 {
-    // TODO: Lisa
     internal static Color s_tintColor = Resources.Colors.Colors.GetColor(ColorName.color_icon_default);
     /// <summary>
     /// Determines if the item is selected.