+++ src/library/DIPS.Mobile.UI/Components/TextFields/Editor/Editor.cs
@@ -6,8 +6,7 @@ public partial class Editor : Microsoft.Maui.Controls.Editor
 {
     public Editor()
     {
-        // TODO: Wait until we have semantic color
-        PlaceholderColor = Colors.GetColor(ColorName.color_neutral_60);
+        PlaceholderColor = Colors.GetColor(ColorName.color_text_placeholder);
         FontFamily = "Body";
         FontSize = 16;
         TextColor = Colors.GetColor(ColorName.color_text_default);